refactor: createPostElement moved to utils (#11)
